backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
playwright==1.48.0
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1
